Renamed `View` to `RelativeViewTransform`
Closes https://github.com/PistonDevelopers/graphics/issues/717
